store: add a mechanism to cache decoded values on a per-column basis
Users still are required to opt-in into caching by using a specific
method `caching_get_ser` and further caching will only work if the
column caching is enabled statically. Dynamic configuration of the cache
size may be possible in the future, but I didn't really want to pull
along config through to this area of the code in order to make this
dynamically configurable.

Any modifications to the cached column:key will cause the key to be
discarded from the cache rather and any future access would have to
fetch it from database anew. This is because the cache stores decoded
values and the transactions store encoded types, so populating the cache
at the write time would require values to be re-decoded even though they
might not be used in the future anyway.

As of this commit no callsites nor any of the column caches have been
enabled yet.