perf: `Arc<BlockHeader>` and `Arc<Tip>`
These data structures are relatively large and yet passed copiously
around by value as well as being cloned a bunch.

While ideally they would be mostly passed around by a reference instead,
I ended up putting them all inside an `Arc` initially because a) this is
less invasive to the code structure; b) after my caching work which
informed this PR in the first place we get `Arc`s out of most operations
pertaining to these types anyway.

In principle the data structure Arcification can be independent and
tested independently of the database changes. I will actually do just
that, but I wanted to post the PR first so that we have something to
look at and discuss if we wanted to.