refactor: return CryptoHash instead of &CryptoHash
This is a follow to a suggestion from
https://github.com/near/nearcore/pull/13653#pullrequestreview-2897372146

Since CryptoHash implements Copy trait it shouldn't be much more
expensive to return it directly instead of the reference which makes it
easier to work with in some cases (for example when having version which
remove certain fields).

Most of the changes I did with the help of cargo fix, manually
removing `*`, and fixing manually some remaining issues. Since I am not
sure how easy this is to review, I am not completely sure how good of an
idea is it to merge this PR. Relying on testing and clippy checks may be
enough.