refactor: return references to CryptoHash from chunk getters.
In some cases for CryptoHash returning a reference may avoid unnecessary
copy of the data so returning a reference would either as efficient or
more efficient as returning a copy from getters. (As far as I understand
it may be as efficient if call will be inlined.)

Additionally using getters with references makes things more
consistent and can highlight some unnecessary copies on the caller's
side.