near-vm + Ditto = near-vm-2 (#13581)
This PR makes a full 1:1 copy of the existing NearVM runtime and then
modifies it to utilize the newly released `finite-wasm 0.6.0`. I deemed
this to be the best way forward to preserve the protocol compatibility
in the light of such an upgrade. The way we handle old protocols now
allows me to remove the old near-vm code after the new code runs in
production.

The intent is to eventually replace this new near-vm-2 into the place of
the original near-vm so that we don't end up dealing with the weird
`-2-` in crate names. I may or may not keep `VMKind::NearVm2`.

This VM is currently not enabled in any way as of this PR -- I am
submitting this as a maximally minimal first step that's as
straightforward to review as possible, so that future functional changes
to this VM can be looked at more easily and in isolation.

FWIW this is exactly an example of where a fully fleshed out limited
replayability scheme would be helpful, but IMO this isn't too bad either
so long as you, as a reviewer, can trust me to not have gone out of my
way to do something nefarious :)

You should be able to verify this by checking that the first commit is
pretty much a result of `cp -r near-vm near-vm-2` and the other commits
are roughly self-explanatory or easily reviewable.