[resharding] Fix check shard_uid_mapping (#13693)
With the new TrieStateResharder, we don't expect the shard_uid_mapping
to exist after resharding is completed for all the newer resharding
events.

However, it's possible for a node to still have the mapping from the
previous resharding events.

With this PR, we are updating the code to remove this check. This was
anyway not a strict requirement and it's fine to not have this check.
It's creating more problems than it's solving.

We had a canary node failing. Context: [#nearone/private > mainnet node
crash during garbage
collection](https://near.zulipchat.com/#narrow/channel/308695-nearone.2Fprivate/topic/mainnet.20node.20crash.20during.20garbage.20collection/with/523037971)

Tested on canary node asrt-mainnet-canary-rpc-01-asia-east1-a-3fa54dd4