[resharding] Do not split state of untracked child shard (#13695)
Related to the thread here: [#nearone/private > mainnet node crash
during garbage collection @
💬](https://near.zulipchat.com/#narrow/channel/308695-nearone.2Fprivate/topic/mainnet.20node.20crash.20during.20garbage.20collection/near/523304892)

High level, during resharding, in the TrieStateResharder, we were
copying the state from parent shard_uid to the child shard_uid even for
shards that the client was not tracking.

This had the unfortunate consequence that the memtrie was removed for
these untracked shards and we were indirectly relying directly on the
state column to iterate through state, which is super slow.

This PR fixes this issue where we are now only copying state in
TrieStateResharder for tracked shards.

For testing, simply adding a debug assert in TrieStateResharder to check
existence of memtrie was failing without the whole set of changes in
this PR.

@staffik, we would need to cherry-pick this into the release as well.