always use constant component in aggregate instruction fees
Having the constant component at 0 would allow one to have a long
sequence of these aggregate instructions with 0 count each costing 0
total gas, which is bad!

Possible optimization would be to instead just add +1 to the count,
which would work too, though not be as flexible (i.e. the fees couldn't
be different for instruction and an incremental step of its execution.)