[shard_tracker] 1/n Add validator_signer to ShardTracker (#13696)
Most of the calls to the shard_tracker pass in `me: Option<&AccountId>`
as a parameter that is always got from the validator_signer account_id.

With this PR we are simplifying the structure and the ShardTracker now
holds an instance of validator_signer which is a mutable config, i.e.
can be changed in the middle of runs. That's why it's not a good idea to
hold a static `AccountId`.

This is the first of many PRs, where later we would use the
validator_signer to fetch the account_id of the client instead of
passing it in as a parameter.