perf: `Arc<Block>` and `Arc<ChunkExtra>` (#13667)
`ChunkExtra` was mostly already wrapped in `Arc` so for that wasn't very
involved, I just enabled some cache for the type. This is based on a
profile where I saw RPC's `process_transaction` reading this data.

Similar reason for `Arc<Block>`: it is a very significant part of
`process_transaction` in RPC actor. That said this is somewhat
opportunistic improvement as technically anybody submitting a
transaction can pick whatever base block they want to use and force the
RPC node to fetch the block from store anyway. But in the happy path
this should make transaction submission about 25% more swift.

But at the same time `Arc<Block>` was quite a bit more involved.

Based on top of #13634