feat(tracing) - better witness distribution spans (#13761)
The `decode_witness` span was a bit wrong - it covered only decoding
witness parts, but didn't include decompressing and deserializing.

Change `decode_witness` to `decode_witness_parts` and add
`decode_state_witness` span which covers decompression and
deserialization. It's not really possible to have separate spans for
decompressing and deserializing because it's done using streaming.

Also add `encoded_witness_size` which tells us the compressed size of a
witness. Now we'll be able to determine what was the compressed and
uncompressed witness size and part sizes using traces alone.