[test] Set reasonable timeout for nayduck tests (#13806)
[#general > nayduck @
💬](https://near.zulipchat.com/#narrow/channel/295302-general/topic/nayduck/near/525948179)

state_sync_massive_validator either completes is 10-12 min or times out
after 1 hour.

This creates the issue where we are unable to retry the test and nayduck
times out after 1 hour. It doesn't make sense to have timeout of an
hour.

This PR changes the timeout to 10 min for most of the tests. This would
ensure we are retrying appropriately.