[refactor] 1/n Introduce chunk iter_new, iter_old (#13783)
This PR is the first step in trying to simplify the chunk iterators. We
currently have iter_raw and iter_deprecated that are not too user
friendly.

With this PR we make some major improvements to the iteration
- Now there are two new functions `iter_new` and `iter_old` to iterate
over new and old chunks
- `ChunkType` now stores a new/old chunk but can dereference to the
chunk header for easy access
- `Chunks` that is the container for the chunk headers can seamlessly
dereference to `[ShardChunkHeader]` so that we can easily use methods
like `len` and `index` on it.

Next PR in the series https://github.com/near/nearcore/pull/13784