[nayduck] CI early exit on test failure (#13808)
In the merge queue, a lot of times we know that a test has failed, yet
the CI workflow waits for all the tests to complete before returning a
failed exit status.

With this PR we do an early exit as soon as we see a failed nayduck
test.

Note that this should still continue to execute ALL the tests in the
nayduck test suite, just that the CI workflow exits early. This is
useful when we want to get a list of all tests tha have failed.

Also note that in the merge queue, if a workflow fails, we `cancel` all
downstream PRs in the merge queue. This would help speed that process up
as well.