implement tx rate controller targeting given BPS
The controller is using the PID loop with the window averaged BPS signal to target the desired value.
Controller runs after the "schedule" part and starts with the TPS of the last scheduled load.
Controller part is optional.
If provided it will try to keep the target BPS and run indefinitely.
If it is missing the `neard` will complete the schedule part and `exit(0)`.
Integral and Derivative gains for the sharded benchmarks (and example file in the `transactions-generator`) are 0.
It is very difficult to pick generally good values for those due to (very non-linear) response of the system.
Reporting the TPS in the `neard` logs is changed to exponentially smoothed instead of averaging its values from the start of experiment.