fix: limit contract deploys part owners (#13849)
Contract deploys message is split into the number of parts equals to the
total number of validators in the epoch excluding validators for the
current shard. We use `reed_solomon_erasure::galois_8` which only
supports max 256 parts.
This PR limits the number of part owners to not exceed the max number of
parts. Technically this should be a protocol upgrade, but attempting to
reed solomon encode data into more than 256 part crashes the node. So
even though during the transition period old binary won't be able to
decode contract deploys from the new one, it is still better than
crashing the chunk producer. The only downside is that the old binary
will log `Failed to reed solomon decode deployed contracts`.