state sync: make concurrency configurable
We have concurrency limits at multiple layers here, so e.g. even if we
have a constant saying that we should be downloading 25 state parts
concurrently, only 6 concurrent tasks would ever get submitted due to
the `buffered` limit.

Even after this change not every instance of this issue is addressed --
for instance the limit of downloads is actually
`external_limit.min(peer_limit)`. Here using multiple task trackers
would be appropriate, but I wanted to keep this PR as self-contained as
it currently is.

The configuration for applies and such may help us to explore more
effective limits than those we have right now without recompiling.