refactor/style: use explicit discriminants for borsh enums
This is a RFC with some example modifications of what the proposed code
looks like. In its essence any time we have an `enum` that's annotated
with either `#[derive(BorshSerialize)]` or `#[derive(BorshDeserialize)]`
we would also require that it would explicitly specify discriminants for
each enum variant and specify `#[borsh(use_discriminants=true)]`.

Motivation for adopting this is that there have been many recurring
instances in the past where a variant would get removed from the middle
of an enum and an unintended protocol incompatibility would occur due to
reordered implicit discriminants.

With explicit discriminants the implicit enumeration of discriminants
gets suppressed, which means that removing variants is no longer as
risky modification to the code as it would've been before. Furthermore,
any messages with legacy, unsupported variants would get dropped during
deserialization rather than making their way into our business code and
handled individually (which also introduces a risk of a mistake.)

If we decide to do this, I intend to explore adding a proper custom lint
to check for this via, possibly, https://github.com/trailofbits/dylint.
Something like dylint would also open the pathway us to lint for other
project style rules as well.