fix: GC epoch change check (#13857)
Currently we update fork tail during garbage collection when epoch
change is detected. The way of detecting epoch change is checking if
epoch id of head's prev block is different from head's epoch id. It was
introduced in #3121. That doesn't always work because GC runs
asynchronously with block processing, so head might move beyond the
first epoch block in between gc runs. This is easily reproducible after
block time reduction on mainnet.

In practice this might result in GC getting stuck. We observed the
following situation with our mainnet node:
- both `tail` and `fork_tail` are on the height `H`
- there is a fork at `H`: blocks at `H+1` and `H+2` are built on top of
`H`, `H+2` is on the canonical chain.
- `gc_stop_height` is much greater than `H`

In this situation `Forks Cleaning` does not happen because
`(stop_height..fork_tail)` is effectively an empty range. `Canonical
Chain Clearing` also doesn't make any progress because
`prev_block_refcount` is 2, so it immediately breaks out of the loop.
The only way for GC to make progress is for the fork tail to be moved
forward, which requires epoch change to be detected.

As a fix we replace epoch change check with change in `gc_stop_height`.
We need to track latest known gc stop height. New `DBCol::BlockMisc` is
added for that.