feat(schema-checker): verify discriminant does not change (#13877)
This covers a situation where an enum with an explicitly specified
discriminant has the discriminant value changed (but the enum otherwise
does not change.)

There are no changes to the protocol types themselves in this PR, the
hashes only change because of the change in how those hashes are
calculated.

This is also in preparation for
https://github.com/near/nearcore/pull/13867 to verify discriminant does
not change from implicit to explicit.