feat(spice): apply chunk asynchronously (#13827)
The main goal of this PR is to move expensive chunk application outside
of actor thread, just like we currently do as part of block processing
and stateless validation.

Implementation notes:
- We send `ExecutorApplyChunksDone` message back to chunk executor actor
so that database changes are happen as part of the actor thread. This is
needed because `ChainStore` is not thread safe. As an alternative we
could create a new instance of `ChainStore` for every chunk application.
- We keep track of blocks that are currently applied with
`blocks_in_execution`. Also we check database for existence of the
corresponding chunk extras to verify if the block was already processed.
- A requirement for previous block to be executed is added in order to
start applying chunks for the current block. Alternatively we could
reject receipt proof for the shards we track and then we would have to
wait until actor saves receipt proof itself to the database.
- Sending messages back to the same actor with `myself_sender` is
monkey-see-monkey-do based on `Client`'s implementation of the same
concept.
- `pending_next_blocks` is changed from `LruCache` to regular `HashMap`
that is actively maintained up to date through removal of the processed
blocks. This is better because no blocks should be dropped, so cache
eviction could hide potential issues. Later we need to add a metric to
monitor the size of the map to make sure it is properly cleared and
doesn't grow in size of time (that would indicate a bug in the
implementation).
- Block reference in the test is changed from `BlockReference::latest()`
to `BlockReference::Finality(near_primitives::types::Finality::Final)`
in order to make it work. With async chunk application query on the
latest block fails because chunks are not applied yet hence chunk extra
is missing. Using final block introduces enough delay to make it work,
but that is just a temporary fix while we look for the proper solution.