state_sync: fix attempt logic for parts (#13891) (#13909)
Similar to #13464, which fixed this logic for headers.

The old code carries the assumption that the fallback source will work
eventually. Rather we should continue making attempts to both sources
until the needed part is obtained.

The fix should be included to the 2.7 release.

---

As a follow up, it should be possible to test this by adding a fake
bucket in the state_sync_decentralized.py pytest. We can also introduce
some random failures of the p2p requests to make sure the bucket gets
hit (unsuccessfully).

Co-authored-by: Saketh Are <saketh.are@gmail.com>