feat: make filesystem contract cache location configurable (#13853)
We have found that it isn't obvious that operators are allowed to clear
out the `~/.near/data/contracts` occasionally, even our own! This led to
this directory ballooning up to 500GB in our snapshots – something that
shouldn't've been in those snapshots in the first place (due to trusting
trust problems these binaries impose onto other nodes – nodes should
regenerate their own binaries always.)

This PR improves the situation in a couple of ways: first it makes the
location configurable and makes the default name `contract.cache`, so
that it is more self-describing as a directory that's not crucial to
node operation. Dumpers and snapshotters can now move this directory to
a location that doesn't get snapshotted by default.

The code will know to rename the old cache directory to the new path in
order to avoid a stale directory with a lot of data in it while the new
one would also grow in size over time.