fix(tests): set vm_kind manually, make x86 wasmtime tests pass
`with_vm_variants` used to take an unused config argument that read from the parameters list
and defaulted to the NearVM case on x86, which caused a misconfiguration when Wasmtime was enabled.

This change removes that unused config argument and defines it inside the closure, as well
as making the `vm_kind` an optional arg for `test_vm_config()` to ensure the correct `vm_kind` is used.

Signed-off-by: miguelraz <miguel@cosmonic.com>
Co-authored-by: Roman Volosatovs <rvolosatovs@users.noreply.github.com>