perf: put Recorder's atomics into their own cache lines (#13936)
Atomics sharing a cache line with other data can evict the shared data
from other caches owned by other cores. That has negative implications
to performance, potentially.

Unfortunately practical profiling does not show an improvement. I see a
fair bit of time spent on a certain atomic operation, but these atomics
are not "it". I still think its worth landing as the change is risk-free
and removes a potential reason to see issues in the future.