[state_part] Make find_state_part_boundary return trie key not nibbles (#13926)
It doesn't really make sense for find_state_part_boundary to be
returning a `Vec<u8>` of nibbles instead of the trie keys as the
convention is to use the trie keys in most places like `iter.seek(..)`
etc.

This PR is required for downstream changes to the parts infra.