refactor: spice: add DbCol to store all next block hashes.
Both for chunk validation and chunk execution in spice we may receive
block before the we have enough endorsements for the result of the
previous one. Without enough endorsements for the result of the previous
block we cannot process any of the following ones.

Endorsements of the block hold no information on which blocks may follow
next so in both cases we need a way to get next blocks to allow
processing related chunks or witnesses.

Using DbCol to store next blocks allows both actors to use it without
additional duplicated logic to track next block on each actor's side.

Part of https://github.com/near/nearcore/issues/13387