perf - extract SW validation into a standalone actor (#13892)
This PR achieves the objective to extract State Witness validation from
the client actor, and places it into a new actor with support for
multithreading.

Normal witness and orphan witness handling are now consolidated a single
place in `chunk_validation_actor.rs`.

The code aim to be functionally equivalent to what we have in master.
The only difference is that ready state witnesses are immediately
dispatched to an actor which is fully dedicated to this operation,
improving latency. The actor is multithreaded so multiple state
witnesses can be pre-validated in parallel.

Simplified architectural change explanation in how State witnesses are
handled.

Before:
```
Networking / SW decoding
...
-> ChunkStateWitnessMessage 
  -> ClientActor: single threaded, handles SW validation, also handles other types of messages
```
With this PR:
```
Networking / SW decoding (no change)
...
-> ChunkStateWitnessMessage 
  -> ChunkValidationActor: multi threaded, handles only SW validation
-> ClientActor: propagates new block notifications, needed to handle orphan witnesses (rare)
```