fix(json-rpc): ensure tx status only skips refunds (#13979) (#13998)
The current implementation for checking the execution status has this
optimization to avoid waiting for refund receipts.

It assumes the last receipt is always a refund. This has never been true
in a general sense but because the protocol almost always had to include
a refund, this bug went under the radar.

Now with the adoption of NEP-536 to reduce the gas refunds, users have
been running into this bug on testnet. See #13872.

The proposed solution in this PR is to check if an outgoing receipt
actually is a refund receipt before excluding it.

## Testing strategy

Run `python3 tests/sanity/rpc_tx_status.py` to check that a tx result
with wait_for="EXECUTED_OPTIMISTIC" includes the receipts for a
cross-contract function call. Run it before the commit with the TX to
see that it previously failed.

Co-authored-by: Jakob Meier <mail@jakobmeier.ch>