Make TokioRuntimeHandle compatible with AsyncSender. (#14006)
This PR implements `CanSend<MessageWithCallback<M, R>>` for
`TokioRuntimeHandle`, acting as a compatibility layer with the current
multisender-based adapter implementation. This allows us to easily
migrate existing actors.

In order to implement `CanSend`, we had to introduce a new trait that
messages must implement. This is because otherwise the `CanSend<M>`
implementation conflicts with the `CanSend<MessageWithCallback<M, R>>`
implementation, as `MessageWithCallback<M, R>` also satisfies `Debug +
Send + 'static`. With a `Message` bound, since `MessageWithCallback` is
not a `Message`, this avoids the conflict. The requirement of the
`Message` trait for handlers is easily taken care of by having
`actix::Message` automatically implement `Message`.