perf(memtries): Avoid re-serializing nodes when size/hash is known (#14009)
Even when the node hash is computed (ie. we already serialized it once),
during computing the new root, we end up calling
`to_raw_trie_node_with_size_non_leaf` in apply_memtries with
`new_with_hash`.

Since we already know the memory usage & hashes for the nodes as well we
can pass it in and avoid the call to
`to_raw_trie_node_with_size_non_leaf`. This could shorten the time of
applying memtries.