test: introduce standardized test loop tooling   (#14017)
This PR introduces `TestLoopNode` to be a standard way of interacting
with the node as part multinode test loop tests. The aim is unifying all
utils we have around interacting with the node in test loop. Examples
are updated as part this PR, I will followup with replacing old utils in
all tests separately.
`account` utils are introduces to simplify the setup state and avoid
`.parse().unwrap()` patter for creating `AccountId` in test loop tests.