fix: replace thread_local in wasmtime-runner (#13972)
This PR replaces `thread_local`s with an `Arc<RefCell<Option<Caller>>>`.
Instead of storing as a `thread_local`, we now access the caller as a
property of the Wasm memory table.

This better suits the use of the caller with `wasmtime`'s API but
doesn't introduce any functional changes. Now that we don't rely on the
thread_local variable, we can begin optimizations for multithreaded
contexts.

As a nice side effect, this also reduced the number of lexical `unsafe`
lines in the codebase.

Co-authored-by: Roman Volosatovs <roman@cosmonic.com>
Co-authored-by: Brooks Townsend <brooks@cosmonic.com>