refactor: parallel transaction processing structure (#13955)
This reworks the way transactions are chunked up and parallellized for
the purposes of transaction validation. This will no longer try to group
or sort transactions by their signer account id – primarily because the
most likely group size is 1 transaction. This is suspected to lead to
non-trivial amount of thread scheduling and work stealing overhead.

Instead we recognize that there are largely just two heavy tasks during
the transaction verification process – signature validation and reading
out accounts/access keys. Transactions are then chunked up into a
reasonably-sized chunks and the chunks for both signature validation and
account/access keys are processed in parallel.

The size of the chunk is chosen to both minimize the risk of significant
tail and to allow some fancy bit-mask tricks to track transaction
validity. At the same time this has made using `ValidatedTransaction`
wrapper type difficult in this code, so you will see a ripple of
`signed_tx` around the code. The way chunk size is computed also ensures
some parallelism in case there aren't many transactions.

The chunk size could be increased to up-to 64 too (and use 64-bit
bitmasks.)

Performance wise this has no siginificant effect on throughput as far as
I can tell, but the code becomes much more straightforward.