feat(wasmtime): enable pooling allocator (#14060)
~This change is based on top of #14057 as it would not do much on it's
own.~

This enables [pooling
allocator](https://docs.wasmtime.dev/examples-fast-instantiation.html#enable-the-pooling-allocator)
to optimize instantiation and minimize memory allocations.

A simple semaphore is implemented, which is used to block in case there
are no free instance and/or table slots

Refs #13791 

~Important thing to keep in mind, is that by default, at most 1000
instances may be active simultaneously, i.e. the runtime with only be
able to handle a 1000 concurrent calls at any point of time, if a 1001st
call would arrive, the instantiation would fail.~

~A potential solution here is to add a semaphore and *block* until
engine has more capacity, but since that would affect performance, this
seems to be undesirable - @nagisa please let me know if that aligns with
the expectations here.~

---------

Signed-off-by: Roman Volosatovs <roman@cosmonic.com>