fix(tokio_runtime): prevent expensive fmts for messages (#14095)
Printing a debug representation every time a message sent not once but
twice is a significant waste of time, especially when messages can be
multiple hundreds of kilobytes long. It also makes it extremely
difficult to follow the logs themselves. Reducing the amount of text can
help, but allocating a string kinda feels… eh.

This keeps formatting of message contents at a an even more verbose log
level (`TRACE`) and makes a point of only printing out the message
contents only once. The "execute"-to-"send" logs can be tied back to
each other using a sequence number of the messages. These are both super
cheap to store and efficient to compute.

Fixes #14085
Fixes #14077

Co-authored-by: Slavas <svyatoslav.savenko@nearone.org>
Co-authored-by: Darioush Jalali <darioush.jalali@nearone.org>