fix: vendor and statically link openssl
This PR adjusts builds to statically link openssl regardless of whether
it is vendored or not, and at the same time uses an indirect feature on
reqwest to request vendoring. Static linking request will continue to
work if we no longer use reqwest though and end up losing vendoring
again.