refactor: spans and messages around partial chunk processing
span being at the debug level fails to record important context for the
warnings that this code can later output. bumping the span up to info
will at least add this context by (usually) default.

using `#[instrument]` is a long time convention at this point so I went
ahead and applied that too.

Redid the warning messages to be self-descriptive and made them use KVs
as per the style convention.