fix: vendor and statically link openssl (#14120)
This PR adjusts builds to statically link openssl regardless of whether
it is vendored or not, and at the same time uses an indirect feature on
reqwest to request vendoring. Static linking request will continue to
work if we no longer use reqwest though and end up losing vendoring
again. And we want static linking of openssl to make sure builds work
even if users don't happen to have a dynamic library conveniently set-up
in the same location as in the builders on the CI. Vendoring helps
reproducibility.

cc #14111