feat(chain): check client/network PV compatibility for Next or NextNext epoch (#14090)
This PR adds the configuration option
`protocol_version_check_config_override`.
It will default to `NextNext`, using the next next epoch PV to determine
when to exit the node.
Validator nodes (that can get a db via state sync), can set this to
`Next` if they desire, allowing them to possibly partake in the last
epoch the client supports.

The prior behavior was to check during block pre-processing, but I moved
it to post-processing to avoid persisting any changes to disk as a
result of possibly finalizing an epoch with an incompatible EpochInfo,
in case an upgrade is missed.

closes: https://github.com/near/nearcore/issues/14052