config: fix computation of the contract cache location
At the time of `Default::default()` the parsing of `store.path`
configuration is not yet done. It seems that this value is filled in
sometime later. Locking in `contract_cache_path` this early thus makes
the value to go out-of-sync with what is specified in the config for
`store.path`.

Instead compute the default value as late as possible.