fix(resharding) - recovery tool for resharding archival data loss (#14185)
This is a recovery tool for the data loss that happened during the
recent reshardings. As the resharding implementation does not store any
TrieChanges, the trie nodes inserted during resharding are never moved
to the cold storage. Once the garbage collection removes those nodes
they are lost and the archival node cannot serve some queries for some
range of blocks after resharding.

This tool replays the resharding, extracts the TrieChanges and stores
them directly into the cold db.

I tested it on my own archival node where I was able to recover the data
and confirm that the reported query that was broken for fast near works
fine.
```
http POST '34.6.25.108:3030' \
     jsonrpc=2.0 \
     id=dontcare \
     method=query \
     params:='{
     "request_type": "call_function",
     "block_id": 160156315,
     "account_id": "astro-stakers.poolv1.near",
     "method_name": "get_reward_fee_fraction",
     "args_base64": ""
 }'
```

The check-only option allows checking if the trie is whole. I need to
work on it a bit more because it's slow and doesn't work if there was a
missing chunk in the first block of the new shard layout.

This is not a fix to the original problem that caused the data loss,
that will follow separately.