chore: more chunk metrics and finer grained buckets (#14115)
The PR introduces 3 new metrics for chunk reconstruction, which will
help detect delays in receiving the needed parts and receipts for a
chunk until it can be reconstructed.

Previously our time buckets were too large which is not always useful,
the PR also introduces `finer_grained_time_buckets` which is used to
obtain finer grained buckets for some recorded metrics.