fix(resharding) - recovery tool for resharding archival data loss (#14185) (#14223)
Cherry-Pick #14185

This is a recovery tool for the data loss that happened during the
recent reshardings. As the resharding implementation does not store any
TrieChanges, the trie nodes inserted during resharding are never moved
to the cold storage. Once the garbage collection removes those nodes
they are lost and the archival node cannot serve some queries for some
range of blocks after resharding.

This tool replays the resharding, extracts the TrieChanges and stores
them directly into the cold db.

The check-only option allows checking if the trie is whole. 

This is not a fix to the original problem that caused the data loss,
that will follow separately.