Refactor check_transaction_validity_period
Checking the validity period consists of two parts:
* Checking that the transaction's block is within X heights of the latest block
* Checking that the transaction's block is an ancestor of the latest block

Currently the two checks are intertwined, let's disentangle them and
have separate logic for each check. There is no change in functionality.
This will help in the future because early transaction preparation
doesn't know the latest block but still wants to check the height
and ancestry using an earlier block.