perf: avoid cloning ApplyChunkResult (#14240)
This PR makes the tradeoff of not cloning `ShardUpdateResult` in
exchange for passing `&mut self` to `get_update_shard_job`.

Alternatively we could "Arc<Mutex<...>>" the cache so it has interior
mutability.
According to recent profiles it takes 10-20ms to clone it, but also I
don't like this clone because it seems like the ShardUpdateResult
doesn't need to be copied.