tx_generator: use final block to submit tx
Use latest final block instead of latest known block
for transactions created by the transaction generator.
With early prepare transactions there can be rare cases where
the preparation job is created when the latest block isn't known,
but it's stuck in the work queue for some time. Then the latest block
becomes known and transaction generator submits a transaction using
this latest block. Only after that the actual preparation job runs,
but it doesn't know about the latest block and rejects the transaction.

This happens rarely and isn't a big issue, but it's more reliable
to use the last final block, which is always known by the preparation job.