perf: config to disable saving untracked partial chunks parts (#14279)
## Reasoning

We don't strictly need to store `PartialChunks` parts if we don't track
a shard in benchmarks.

- Parts are in the memory cache anyway (for a while); CP keeps them for
3 epochs
- This reduces write pressure by a good 20%
- Leaked `PartialChunks` are less "impactful"

## Changes

- Add config option to disable persisting chunk parts, active only in
benchmarks

## Caveats

It turns out we still need receipts and `PartialChunks` header .. so the
workaround is to avoid storing the parts which are the biggest
contributor to size.

## Results

This makes the GC+compaction combo work as expected. Also less write
pressure in general on the disk

<img width="634" height="339" alt="image"
src="https://github.com/user-attachments/assets/6806d204-f35d-4446-850b-1bdd723d536d"
/>