Estimate storage proof size when reading from TrieUpdate
Early transaction preparation will read from the post-state
TrieUpdate, which means that some of the entries will not be
in the Trie, only in the TrieUpdate's key-value hashmap.

This makes it harder to correctly estimate the size of storage proof
generated while reading data needed for transaction validation.
Reading from TrieUpdate doesn't record anything in the recorder,
so relying on Trie recorder's storage proof size counter is not
enough. We could have situations where reading from the TrieUpdate
is cheap, but the actual Trie read that nodes applying the chunk
will do generates a lot of storage proof (as much as 32KiB per read!).

TrieUpdateWitnessSizeWrapper helps estimate how much storage proof the
reads from Trie will actually generate and keep the size of storage proof
reasonable.