Prepare transaction jobs in ChunkProducer
After receiving PostStateReadyMessage the chunk producer
will start a background job to prepare the transactions.

Later when it comes time to produce a chunk, the chunk producer
will wait for the job and take the prepared transactions.

If the job doesn't exist or was started with the wrong inputs,
the job result is discarded and preparing transactions happens
normally.

This commit includes the code for starting the jobs, managing them,
cancelling and waiting for the results.

Actually handling the PostStateReady message will be done in a separate commit.