feat: host function `promise_set_refund_to` (#14285)
With `promise_set_refund_to`, an outgoing action receipt can set where
the balance refund should go, in case of a balance refund. Balance
refunds happen after a failed receipt with attached deposits.

Adding the host function is part of of NEP-616.
(https://github.com/near/NEPs/pull/616).

It helps with cases where the signing account calls a deterministic
account, which then calls another deterministic account. If the second
call needs a balance attached, the signing account may have provided it
for forwarding. In that case, if the second calls fails, we want the
funds to go back to the signer, not the caller.

The implementation is involved. It requires a new field somewhere on the
receipt. I think putting it on the ActionReceipt makes the most sense.
This is the first time we modify `ActionReceipt` in NEAR history, as far
as I know.

I propose to make ActionReceipt versioned, by adding more variants to
`ReceiptEnum`. The variant implicitly defines the version. Note that
this needs to be done for `ReceiptEnum::Action` and
`ReceiptEnum::PromiseYield` since both contain an action receipt.

To make changes around the code base easier, I then added `pub enum
VersionedActionReceipt<'a>` which can be constructed from either
`ActionReceipt` or `ActionReceiptV2` and has access methods to all
fields.

I also considered making `ActionReceipt` itself versioned instead of
expanding `ReceiptEnum`. But for backwards compatibility with borsh
encoding, I believe it would still need new variants in `ReceiptEnum`.

We could only go to `ReceiptEnum::VersionedAction` instead of
`ReceiptEnum::ActionV2`. But I don't see how it would be better than my
proposed solution.

# Testing strategy

Tests check that the new host function can be called from a contract and
correctly redirects the refund, if the feature is enabled.