[actix migration] Migrate network off of actix. (#14280)
This migration moves `PeerManagerActor`, `PeerActor`, `NetworkState` and
surrounding structures away from actix into the near-async actor
framework. This is the last non-test usage of actix.

Summarizing changes from top down (in the order displayed by github):
* Test utils, such as `chain/network/src/actix.rs`, are massively
simplified to use the near-async framework directly as much as possible.
* `Demux` is changed to use `FutureSpawner` instead of directly spawning
to tokio. It means that the demux tasks run directly on the actor which
owns it.
* `PeerActor` migration:
* Rather than using the `actix::Arbiter` (for spawning futures) and
taking in the `actix::Context` (for spawning futures, sending messages
to self, and running delayed actions) all the time, we have the
`PeerActor` just keep a `handle` to itself, allowing it to perform all
of these directly. Most of the changes in this file is dedicated to
switching ctx calls to handle calls.
* Note that `ctx.spawn(wrap_future(...))` is an actix-specific mechanism
to spawn futures allow chaining with a function that takes a `&mut
Self`. This has been replaced by `spawn` followed by `run_later` with a
delay of zero. It has the same effect.
* `FramedStream` is simplified to take a `Sender<Frame>` and
`Sender<Error>` instead of the whole actor addr/handle. So this type is
no longer generic.
* `PeerActor` has a special
`registration_buffered_actions`/`delay_if_registering()` mechanism as a
replacement to the old `ctx.wait(registration future)`. The reason is
because the peer actor is not programmed to properly handle any messages
(even disconnection) while the handshake is still processing; the
processing of handshake is async, thus the need to delay other messages.
The old `ctx.wait` accomplishes this, but this is a very ugly thing to
support in the actor framework, so instead we delay the messages
manually if a registration is happening - that's nice too because it
explicitly documents the behavior.
* We no longer ask the other peer to remove us from their store, if the
disconnection is requested by the PeerManagerActor. I don't know why it
was the case before, but we actually test that this doesn't happen in
`test_storage_after_disconnect`, except that the test was relying on the
subtle timing before that disconnection seemed to always result in a
broken stream (the Disconnect message is never fully sent). But now
somehow the Disconnect message *may* actually be fully received, so this
test is flaky otherwise.
* FramedStream, like mentioned above, is refactored to just take two
senders and a future spawner.
* `Connection` and `NetworkState` undergoes the similar change of using
`TokioRuntimeHandle` and `FutureSpawner`.
* In `NetworkState`, there is a `Runtime` field, coming from the obscure
concurrency library in the network crate. For now, I'm leaving it
untouched, but in the next PR this will also be migrated to the
near-async tokio runtime. For now also, I'm changing the `actix::spawn`
to `self.spawn`.
* The `PeerManagerActor` migration is very similar to `PeerActor`'s but
without the added complexity of buffering messages during registration.
Not much to say here.
* In `snapshot_hosts.rs` there is a compilation subtlety that because
peer manager actors can handle messages with both a return value of ()
and a response, we need to explicitly specify the return type.
* In `raw/tests.rs`, we reveal a subtle difference between actix and our
own framework: our actor framework does not guarantee order of execution
of futures being `spawn`ed. In fact, there should be nothing that
actually relies on such order. This test, `test_raw_conn_state_parts`
assumes such ordering, so I've changed the test to no longer assume
that. In production, StateResponse is identified by the part ID among
other things so we should never need the messages to be ordered.
* Remove most of the `actix` functionality from near-async.
* Migrate a bunch of tests to no longer use `run_actix`, which mostly
involves directly making them `tokio::test` and just running directly
without the `run_actix` wrapper, and ensuring that we construct an
`ActorSystem` and tear it down after.
* Make `neard` no longer run an actix System since we have no production
actix actors anymore. Bootstrap async context using a single-thread
tokio runtime.
* Migrate other tools to no longer spawn actix system as well.
* Remove `near_performance_metrics::actix::spawn`.