chore(cloud_archival): Differentiate between cold and cloud archivals (#14305)
**Context**
With the introduction of cloud archival nodes, the meaning of the
`archive` config option has become ambiguous.
Previously, archive referred either to split storage or to a legacy
archival node.
Now, it can also represent a cloud archival reader, writer, or both.

**Changes**
* Clarified the distinction between cold and cloud archival in the
codebase, adding TODOs to revisit after MVP.
* Enabled independent creation of cloud archival (separate from cold
archival) in the testloop.