feat: spice: make data distributor errors more specific.
This PR adds allows distiguishing errors when receiving data which makes
it easier to make sure that tests fail in the expected branch (when
either processing incoming data with or without the relevant block).

Distinguishing when data is already known errors allows us to reduce
noise of logs when investigating issues.

Part of https://github.com/near/nearcore/issues/13611