refactor: Apply new chunks without chunk header. (#14318)
There are several reasons to do this:

- For new chunk application we only need some of the fields from the
chunk header so there is no need to clone and pass-in it whole.
- With spice we will want to treat missing chunks as empty chunks. In
that case there would be no chunk_header to use for application and it
wouldn't make sense to create a dummy one for that purpose.
- Spice chunk header doesn't include information needed for application
this so PR moves us closer to removing
`into_spice_chunk_execution_header` function, which is technical debt at
this point.

Part of https://github.com/near/nearcore/issues/14300