feat: spice: Use separate struct for spice endorsements.
We will implement execution of shards for spice even when chunks are
missing so it makes sense to make endorsements consistent with witnesses
and use the same chunk_id to key them. Because of the differences and
completely different code-paths it makes the most sense to introduce
new types for spice endorsements.

As part of addding new endorsements I added their propagation through
the network and tested it with localnet.

New keying of endorsements also means that we would have to validate it
differently which is also included in this PR.

Part of https://github.com/near/nearcore/issues/14300