wasmtime: include machine information in contract cache hashes (#14332)
wasmtime can determine host's capabilities and take advantage of the
information much more flexibly compared to NearVM. If we want the cache
to work correctly, however, we must make sure that we don't use binaries
compiled for a different host, as that can lead to precompiled binaries'
failure to load or – in case of oversights in compatibility checking
code
– crashes.

Surprising to me was that we nowadays use the cache information to drive
some business logic and not just as a plain cache. In particular these
cache contents are used to determine whether node would request a
contract from another node or not. This results in some interesting
interactions. For example, even if wasmtime ends up detecting
that the precompiled code is not compatible at instantiation time, we
won't have the wasm code to compile and handle that error on-the-fly.
We must be pretty sure that we'll be able to load the contract when
the partial witness actor does its work. The only way to tell this 
with any sort of reliability is to just hash everything about
host and code that might influence the code generation.

This however also means that the cache reuse is not as potent as it
was with near-vm. For instance, sharing the cache between multiple
machines will be more likely to not result in any re-use now.

cc @rvolosatovs