feat: promise_batch_action_state_init (#14364)
Add a new host functions to create Action::DeterministicStateInit on
chain.

- promise_batch_action_state_init
- promise_batch_action_state_init_by_account_id
- set_state_init_data_entry

This is a deviation from the [original reference
implementation](https://github.com/near/nearcore/pull/14080)
where there was a single host function cover all of this.

We discussed that we want to avoid making borsh serialization
a part of the guest-host interface in a
[draft
PR](https://github.com/jakmeier/nearcore/pull/4#discussion_r2382325125).