test: find trie split (big random trie) (#14323)
A test which verifies that the best split account is found, using a big,
randomized trie. A couple of issues were uncovered with this test and
fixed:

1. The found path needs to have an even number of nibbles, and a minimum
length of 4 to be parsed as an account ID. To address this,
`force_next_step` procedure was added.
2. The condition in `find_middle_child` should be `>` rather than `>=`.
The 'equal' case represents a perfect split – in such case we should
pick the next child to cut off exactly half of the memory.

I've run the test on many seeds and picked the problematic ones to
include them in the final test code.