fix(det-accounts): store deposit in pre-initialization (#14388)
It's an explicit feature in NEP-616 to allow sending a Transfer to a
deterministic account id first, already creating the account but not
allowing it to be called, yet.

This fixes a crucial bug in the pre-initialization creation, where the
deposit was not deposit and just vanished instead.

`test_deterministic_state_init_prepay_for_storage` checks that the
described feature works as expected and that the sent funds arrive at
the receiver.