feat: spice: Sign partial data messages. (#14359)
Signature is required to make that sender in the message is correct.

As part of this PR I had to move messages to near-primitives to allow
the use of signature differentiator, and adjust some error messages to
avoid loosing some of the logs usefulness with more restricted API on
message with unverified signature.

Part of https://github.com/near/nearcore/issues/13611