Replace network/src/concurrency/runtime.rs with our actor framework (#14391)
This runtime is adhoc and not tracked by the actor framework, so we
migrate it so it can be tracked.

Also, there are some cases such as network graph update computation that
should belong to a synchronous thread instead of a tokio thread.

To keep the semantic of owned worker threads, we introduce the concept
of *owned* actor runtimes. These are the same as those spawned via an
ActorSystem, except that the lifetime of these runtimes are owned by the
handle(s). Dropping the handle causes the runtime to be stopped. This is
important because the network graph update code assumes that the worker
thread is still alive when waiting for results from it.

This un-parallelizes the graph update computation. This is because
somehow the rayon parallelization causes the thread to be stuck (that
rayon library in concurrency should probably also be replaced since we
don't want untracked work going on in rayon threads either, but that's
for later). I verified in a test mainnet node that this computation is
very short (on the order of 1ms without parallelization) so we shouldn't
really need to parallelize.