feat(cloud_archival): Persist per-block blobs in external storage (#14421)
Enables saving block data to external (cloud) storage.

Changes:
- Spawn `CloudArchivalWriter` using `FutureSpawner` instead of running
it as an actor. This was required because `ExternalConnection` uses
async calls. As a side effect, `initialize_cloud_head` had to be moved
into the future, so an error there will no longer prevent `neard` from
starting up.
- Introduce the versioned `BlockData` struct that aggregates
block-indexed data for archival.
- Add `CloudStorageFileID` to manage file paths for cloud archival data.
- Implement `CloudStorage` that uses `ExternalConnection` to upload and
download data. For the MVP, only the local filesystem and `GCP` backends
are supported; `S3` is disabled for now.
- Unignore `test_cloud_archival_resume` since the cloud head is now
persisted to external storage.