fix: spice: Write all core statements from a separate actor. (#14440)
Before this PR endorsements can be recorded either separately or as part
of blocks. Additionally execution result is recorded only after enough
endorsements are present. We depend on this when producing core
statements for blocks - we include in core statements all endorsements
and execution results we have in store. (Since we write execution
results only after endorsements, we know that all execution results in
store are endorsed.)

Because endorsements and execution results are stored in separate db
columns there is a place for a race condition between generating core
statements and recording endorsements (or more broadly there's a
possibility of breaking an invariant that iff enough endorsements exists
we have execution result in store). This can happen when we are
processing 2 endorsements at the same time and both of them together are
enough for endorsing of the execution results, but each on it's own
isn't.

Since racing endorsements can come at the same time when recording a
block and a standalone endorsement, it wouldn't be enough to use mutex
for writing endorsements. (During block recording endorsements and
execution results from block are written later on.)

Another alternative I considered is to evaluate if execution results are
endorsed during reading (and not writing), however when testing with
localnet it resulted in execution moving significantly slower than block
production (even with empty blocks).

Using a separate agent to write all endorsements and execution results
we make sure that those columns contents are consistent and we have
execution result iff enough endorsements are present. (It would be
possible to achieve the same by writing core statements from block after
block processing is done and using mutex, but it's much easier to
introduce other race conditions with this approach compared to using a
separate actor.)

Note that since endorsements and execution results are now written by a
separate actor we cannot rely on those fields during block validation.
To make it possibly to verify if core statements in blocks are valid I
hence added an additional field to uncertified chunks info containing
all endorsements that known - using it we are able to evaluate if
execution result is expected in the block without reading endorsements
and execution results from the store.

An added benefit of separation is that it becomes much easier to read
core statements and there is no longer need to have core reader passed
around to make sure that it's the same between all users.

To make it easier to review I copied for implementation and tests old
spice_core.rs to spice_core_writer_actor.rs in a separate commit
([7ce20c5](https://github.com/near/nearcore/pull/14440/commits/7ce20c52a304f6ac1f1f819defdd30b88dc04888))
so that diffs for both can be compared from the point of view of what is
being removed.

Part of https://github.com/near/nearcore/issues/13388