perf: remove superfluous spans and improve metrics
in a hot loop like what we're seeing when applying many very cheap
receipts the span handling is actually becoming quite expensive
(although still not as expensive as the metric's with_label_values.

If per-action spans are useful for investigation in the future, they
should be added for those actions independently. Every other layer can
be inferred (there's very little that's happening between
`process_receipt_with_metrics` and `apply_action`.

Remove or reduce the cost of these operations to a minimum.