perf: allow `take_result` to proceed before transaction reintro
`reintroduce_transactions` take a non-trivial amount of time to complete
(though not quite as much as the transaction preparation itself, the
ratio being roughly 90:10.) There is no reason to block taking of the
result once it has been computed. This also gives less leeway for
a stray `cancel` to mess things up.