[sync] rework external_backoff (#14204)
The way the backoff timers were implemented previously was highly
suboptimal.

The intention was that if we fail to download a part from the cloud
bucket, we should wait a while before making the same request. That is
because a failure from cloud indicates that either the cloud service is
misconfigured/down/etc. or the requested file has not been uploaded yet.
In either case it does not make sense to rapidly re-attempt download
requests, so a "backoff" of one minute is quite reasonable. Without it
we have seen nodes spam thousands of failed download attempts in
minutes.

However, the way it was implemented was by blocking one of the state
sync "threads." We strictly limit the number of concurrent requests and
it does not make sense for one failed cloud request to artificially
block all progress.

This PR implements the intended fix in a better way without changing the
user-facing config parameters.

#14020