fix(indexer): expose local receipts separately from chunk receipts (#14424)
Chunk receipts represent outgoing receipt from application of the
previous chunk. So local receipts from the current chunk clearly don't
belong there.
This is breaking change and we need to support client that happened to
rely on local receipts being part of the chunk. New field
`local_receipts` is added to cover that.

Part of #14109