feat: improve indexer local delayed receipt lookup (#14475)
This PR contains improvements around logic for local delayed receipt
lookup in the previous blocks:
- Don't fetch receipts for outcomes, we don't need those
- `transactions.into_iter().zip(outcomes.into_iter())` is incorrect
since failed transactions can generate no outcomes
(`ProtocolFeature::InvalidTxGenerateOutcomes` is not enabled yet).
Instead find outcome that generated to receipt we are looking for and
then find the corresponding transactions.
- We know shard_id receipt we are looking for belongs to, so we can
limit our lookup to only transactions in the shard, but then we'd have
to take resharding into account. This is complicated and hard to test,
so we instead we chunk check all shards as in the original
implementation.
- Generic refactoring to make code look better

Part of #14109.