fix: increase upgradable test epoch length (#14502)
In one of the failed test runs the reason was that one of the validators
was kicked out:
```
'prev_epoch_kickout': [{'account_id': 'test3', 'reason': {'NotEnoughBlocks': {'expected': 16, 'produced': 3}}}]}
```

When looking into it I found the following in the validator logs:
```
Skipping block productionheight=122 err=BlockProducer("Should reschedule")
client: Skipping block production, prev block is not caught up height=122
```

It looks like state sync didn't manage to finish in time.

Let's try increasing epoch length and see if it helps. This is not a
great solution as it makes the test ~1.5x slower, but let's see if it
helps. If confirmed I will look into making the test faster.