fix(spice): disable parts of indexer for spice.
Some, but not all, indexer functionality is required in order for
mirroring of mainnet traffic in forknet to work.

We don't yet write to the correct columns in a way indexer assumes.
Adding writing to Receipts column isn't enough to make it work.

Part of https://github.com/near/nearcore/issues/13618