feat(spice): Persist data required for distribution. (#14552)
In this PR I add persistence of the witnesses and getting the data from
the store instead of from the cache when requested for distribution.

I also made changes to make sure that we store in memory only the
required data, namely ids and parts for data that we are waiting on.
This later can be populate on restarts.

Out of scope of this PR:

* population of required data on restarts,
* stopping requesting witnesses past final certification head,
* gc-ing of witnesses older than final certification head.

Part of #14485