pytest: delete Proxy and PeerMessage libs (#14545)
The PeerMessage and Proxy libs re-implement the network protocol spec in
borsh format to enable tests which create or modify network messages
between nodes. We can achieve the same thing with TestLoop now.

Also, we have long since deprecated borsh in favor of protobuf for the
network message protocol. This PR unblocks cleanup of the borsh code
(#13412).

The following trivial tests are simply deleted:
- proxy_simple.py
- proxy_example.py
- proxy_restart.py

Two non-trivial tests are ported to TestLoop:
- block_chunk_signature.py
- network_drop_package.py

The following tests are deleted and we will decide whether to port them
as a followup:
- sync_ban.py
- sync_chunks_from_archival.py
- repro_2916.py
- peers_request.py

This will also unblock deletion of the deprecated AnnounceAccounts,
which we couldn't do because the new AccountsData is not supported by
the borsh-only proxy lib.