feat(spice): disable parts of indexer (#14530)
Some, but not all, indexer functionality is required in order for
mirroring of mainnet traffic in forknet to work.

We don't yet write to the correct columns in a way indexer assumes.
Adding writing to Receipts column isn't enough to make it work so for
now I disabled parts of the indexer for spice so that forknet can be run
from master. Later we can figure out proper implementation for indexer
with spice as part of https://github.com/near/nearcore/issues/14492

I tested that this PR is enough to make forknet traffic mirroring run.

Part of https://github.com/near/nearcore/issues/13618