fix(mirror): remove assumption that no channel can be closed
When final select in `run` function finishes receivers of some senders
there are passed into spawned coroutines would be deallocated. This
would cause send on those senders to return error. Since we were
unwrapping before this change it would cause panic. This panic would
show that some values weren't able to be send, but have no information
about what is causing that - which is a reason why run finished and
deallocated relevant receivers.

If we instead don't unwrap on send it should give opportunity for the
run function result propagating and hopefully showing relevant
result.

Note that I haven't tested this change since it's not clear how to
trigger some failures reliably.