test: allow TestLoopNode own NodeExecutionData
This change makes it easier writing tests when we need to create TestLoopNode and
then need mut reference to test loop env. A good example of this is
restart_node example test that I updated as part of this PR.

Note that before this change there was a subtle mistake in a test:
```
let restarted_node = TestLoopNode::from(&env.node_datas[validator_index_to_restart]);
```

Would create TestLoopNode that references node datas of the killed node,
but not of the new restarted node so althought for this test it doesn't
matter (since storage is shared) using senders, I believe, would have failed.