feat(spice): allow nodes to recover from restarts
We do not need to do anything for witness validator actor since it would
get triggered after we request and receive a new witness.

As part of this PR I added the test for an existing behaviour (final
execution head updating when we aren't tracking any shards) since
distributor actor now depends on it.

Part of https://github.com/near/nearcore/issues/14485