feat(spice): allow nodes to recover from restarts (#14582)
We do not need to do anything for witness validator actor since it would
get triggered after we request and receive a new witness.

As part of this PR I added the test for an existing behaviour (final
execution head updating when we aren't tracking any shards) since
distributor actor now depends on it.

Closes https://github.com/near/nearcore/issues/14485

---------

Co-authored-by: Darioush Jalali <darioush.jalali@gmail.com>