chore(async): log msg execution as trace (#14585)
I don't think we need to log each message's execution as debug which
will have it logged in CI as is.
Perhaps it will improve flakiness of some pytests.

Also tried to apply some of the newer logging guidelines. Not sure if
further changes are needed, but I think this is an improvement.