fix(spice): request witness if we apply current block.
We may not need to request receipts from peers if we apply previous
block ourselves.

We don't need to request witnesses only if we apply the block witness is
for. Before this PR the code mistakenly was making decisions whether we
need a witness based on if we apply previous block, not current. If we
weren't applying a prev block and are applying block, then we may
unnecessarily request witness, and symmetrically we may fail to request
witness.

Similarly when deciding to request receipts we only need receipts for
shards that we apply in current, not previous block.

This particular behaviour is hard to unit-test so I omitted the tests.
Generally scenarios like these we would need to handle and test
carefully when implementing proper epoch transition handling with spice.

Generally this can be fixed later, but somehow I encountered this in
forknet.