feat(near-tracing): stream body in /raw_trace (#14581)
Currently when handling a `/raw_trace` query, the tracing server reads
the whole response into memory and then returns it. This causes problems
when the response is big - with 60 shards a 1 minute of traces takes
over 60GB of memory, which can easily cause OOM on the tracing server.

To deal with this let's stream the response - immediately output each
span right after reading it from the database, without reading the whole
response into memory. This should keep memory usage minimal.

~~I wanted to do something similar to Python's async generators which
yield subsequent parts of the response, but sadly Rust's async doesn't
have generators yet. I wrote a helper struct to make this style of
output generation work despite the lack of language support.~~