fix(sharded_bm.py): fix setcap on a local binary (#14633)
When a local file path is provided as binary url, `neard-runner` creates
a symlink to the binary file and puts the symlink in
`neard-runner/binaries`.
The problem is that we can't run `setcap` on a symlink, it has to be run
on a real file. Trying to run it on a symlink causes an error:
```bash
$ sudo setcap cap_sys_nice+ep ~/.near/neard-runner/binaries/neard*
Failed to set capabilities on file `/home/ubuntu/.near/neard-runner/binaries/neard0' (Invalid argument)
The value of the capability argument is not permitted for a file. Or the file is not a regular (non-symlink) file
```

Let's fix the issue by running `setcap` on the actual file.