Remove required from config structs in openapi.json (#14621)
Some of the configs change over time, which breaks generated jsonrpc
clients.

Namely, `ProtocolConfig` and `ClientConfig`.
They have enpoints in json rpc, under `/EXPERIMENTAL_protocol_config`
and `/client_config`. When a new field is added to one of the stucts, a
newly-generated client will expect that field to be in the response.
This causes the client to not accept responses from the blockchain nodes
with older nearcore version.

This PR removes `"required"` field in openapi.json for `ClientConfig`,
`ProtocolConfig` as well as some of their subfields.

Though it doesn't cover cases when some new subconfig is added to i.e.
`ClientConfig` without handling `"required"` for it, and it's updated
later. Therefore, `"required"` field should be removed for all the added
subconfigs on time.


The list of configs to remove "required" field for now:
ClientConfig
 - GCConfig
 - CloudArchivalWriterConfig
 - StateSyncConfig
   - DumpConfig
   - ExternalStorageConfig
   - SyncConcurrency
 - EpochSyncConfig
 - ChunkDistributionNetworkConfig
 - ChunkDistributionUris

RpcProtocolConfigResponse
 - RuntimeConfigView
   - RuntimeFeesConfigView
     - DataReceiptCreationConfigView
     - ActionCreationConfigView
     - StorageUsageConfigView
   - VMConfigView
     - LimitConfig
     - ExtCostsConfigView
   - AccountCreationConfigView
   - CongestionControlConfigView
   - WitnessConfigView