feat(spice): run all passing tests with spice
This change allows running with spice all but failing tests, which are
ignored for now. Eventually most tests should be fixed as
spice feature matures, but this setup with separate run of the tests
with spice allows us to make sure that all tests pass both with and
without spice which is important for nightly features that are released
before spice. It also would allow us to gradually fix tests failing with
spice one by one.

The only possible downside is that it may be slightly expensive running
both with nightly and nightly with spice separately. Let me know if you
have any suggestions for a better approach.