perf: stabilize 64 shards benchmark (#14666)
This PR addresses two source of instability in the 64 shards benchmark:
1. `transaction generator` settings: now the generator has a stable
schedule and a good block time target to reach 1M TPS.
2. Fixed issue with chunk availability after `init` command. In the
workflow of `sharded_bm` we start the chain twice. This creates an issue
because we don't persist chunks in DB for all nodes and we stop all
`neard` at the same time. It can happen that a CP is missing a chunk for
its own shard (because it was produced by the other CP) and the part
owners nodes are stopped before they can answer the chunk request; thus,
upon restart the node is stuck with a missing chunk that potentially
nobody has, if the other CP is not a part owner or not the target of the
request. The fix is to persist chunks on all nodes during the `init`
phase.

Open issues:
- tx generator fills the tx pool too quickly because it's not aware of
other CP creating transactions. Not a blocker, the chain works even with
the full pool
- after 30 minutes nodes start to go OOM
- performance not confirmed with latest master neard

[grafana](https://grafana.nearone.org/goto/cf4sy58zlam80e?orgId=1)