feat(spice): Introduce spice protocol feature and related version. (#14688)
It would take awhile to release spice. In the meantime the features we
would release should be tested and work without spice. In particular
because spice change changes a log of usual things and disables some
features (like optimistic block) testing feature with spice would not
guarantee that the feature would work without spice. Having a separate
protocol version for spice depending on spice cargo feature allows us to
achieve this.

Once spice is close to release we should make it part of nightly and
cargo feature should be removed.

Refs https://github.com/near/nearcore/issues/14489