refactor(spice): use block headers in core apis.
The main reason we couldn't use headers before is because we needed
block to build genesis chunk extra. With this refactoring genesis chunk
extra can be build without a block though now we need to store genesis
gas limit in spice core reader.

Using a header in API instead of blocks would allow using it in places
where we already get header without the need to request the block as
well. For example when preprocessing the block here:
https://github.com/near/nearcore/blob/db9ce5b21f2f8a9ee235acaa442a165c775f90b9/chain/chain/src/chain.rs#L2313