feat(client): improve the metric about transaction preparation limits (#14716)
We have a metric which shows us what kind of limit was hit when
preparing transactions (gas, size, storage proof etc).
But the metric doesn't record anything when preparation didn't hit any
of the limits and simply ran out of transactions to pick from the pool.
This can be confusing, sometimes I see zero data in Grafana and I'm not
sure whether no limits were hit or something broke with the metrics.
Let's also record when the preparation was limited by running out of
transactions in the pool, this will make it easier to reason about the
data. It will also make it consistent with another similar metric that
we have - `CHUNK_RECEIPTS_LIMITED_BY`.

Apart from that I cleaned up a little bit to make things nicer and
renamed the metric, as the previous name was unwieldy.