fix(network): move PartialEncodedChunkForward to T2 (#14739)
`PartialEncodedChunkForward` was moved to T1 as part of #13704, but such
change is not backward compatible. Node on the old binary version does
not expect this type of message to be sent over T1 and it will [close
the
connection](https://github.com/near/nearcore/blob/2.9.1/chain/network/src/peer/peer_actor.rs#L1738).
This results in important messages sent over T2 which caused significant
drop in chunk endorsement rate in mainnet.

We keep enum variant in T1 as `_PartialEncodedChunkForward` to make
future migration easier.